% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Generator.R
\name{get_playlist}
\alias{get_playlist}
\title{Playlist Generator}
\usage{
get_playlist(genre, nsongs)
}
\arguments{
\item{genre}{the genre}

\item{nsongs}{the number of songs in a}
}
\value{
An object of the song class
}
\description{
Returns a playlist for a user given a genre other characteristics
}
