% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{new_playlist}
\alias{new_playlist}
\title{Playlist Constructor}
\usage{
new_playlist(songs, genre)
}
\arguments{
\item{songs}{list of song objects}

\item{genre}{a character representing what genre the songs in the playlist all fall under}
}
\description{
Constructor for Playlist class
}
