% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{new_song}
\alias{new_song}
\title{Song Constructor}
\usage{
new_song(title, author, duration, genre, danceability, energy, key, tempo)
}
\arguments{
\item{title}{Title of the song}

\item{author}{Author of the song}

\item{duration}{The length of song in seconds}

\item{genre}{One Genre of the song (spotify can classify songs in multiple
genres - in this case whatever the user has asked for is the genre)}

\item{danceability}{On a scale of 0 to 1}

\item{energy}{Measure from 0 to 1 on "perceptual measure of intensity and
activity" where 1 is more active}

\item{key}{Major or minor, C - B}

\item{tempo}{in BPM}
}
\description{
Constructor for song class
}
