% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{playlist}
\alias{playlist}
\title{Playlist Helper}
\usage{
playlist(songs, genre)
}
\arguments{
\item{songs}{A Vector of the type Song}

\item{genre}{Genre of the playlist}
}
\description{
The helper function called in playlist generator to make playlist
}
