% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{song}
\alias{song}
\title{Song Helper}
\usage{
song(title, author, duration, genre, danceability, energy, key, mode, tempo)
}
\arguments{
\item{title}{Title of the song}

\item{author}{Author of the song}

\item{duration}{The length of song in milliseconds, converted to seconds}

\item{genre}{One Genre of the song (spotify can classify songs in multiple
genres - in this case whatever the user has asked for is the genre)}

\item{danceability}{On a scale of 0 to 1}

\item{energy}{Measure from 0 to 1 on "perceptual measure of intensity and
activity" where 1 is more active}

\item{key}{Number 1 - 12 that is translated into Words}

\item{mode}{Major (1) or Minor (0) combined with key to produce the key of song in character value}

\item{tempo}{in BPM}
}
\description{
The helper function called in song generator that makes a song
}
