% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{summary.playlist}
\alias{summary.playlist}
\title{Summarizing Songs in the playlist}
\usage{
\method{summary}{playlist}(object, ...)
}
\arguments{
\item{object}{An object of the type playlist to be summarized}

\item{...}{to adhere to generic summary method}
}
\description{
Summarizes each song & it's characteristics
}
