% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{summary.song}
\alias{summary.song}
\title{Song Summary}
\usage{
\method{summary}{song}(object, ...)
}
\arguments{
\item{object}{an object of type song to summarize in more detail}

\item{...}{to adhere to generic summary method}
}
\value{
A data frame including the Title, Genre, Author, Danceability,
Energy, Key and Tempo of each song in the playlist.
}
\description{
Summarizes a song & it's characteristics in the form of a
data frame to be easily combined with other song data.
}
