% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{validate_playlist}
\alias{validate_playlist}
\title{Playlist Validator}
\usage{
validate_playlist(obj)
}
\arguments{
\item{obj}{An object of the type playlist}
}
\description{
Makes sure an object with playlist has the correct attributes of playlist
}
