% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{validate_song}
\alias{validate_song}
\title{Song Validator}
\usage{
validate_song(obj)
}
\arguments{
\item{obj}{An object of the type song}
}
\description{
Makes sure an object with song has the correct properties of a song
}
