% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{print.song}
\alias{print.song}
\title{Print Song}
\usage{
\method{print}{song}(x)
}
\arguments{
\item{song}{An object of type song to print out}
}
\description{
Prints objects of song class in neat format
}
