% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Generator.R
\name{rec_playlist}
\alias{rec_playlist}
\title{Playlist Generator}
\usage{
rec_playlist(genre, nsongs, ...)
}
\arguments{
\item{genre}{the genre a user wants all songs in their playlists to fit under}

\item{nsongs}{the number of songs in the playlist, maximum of 15}

\item{...}{Other specifications for the playlist, all specifications for rec_song welcome}
}
\value{
An object of the song class
}
\description{
Returns a playlist for a user given a genre other characteristics
}
