% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reccomendor.R
\name{rec_song}
\alias{rec_song}
\title{Recommend a song}
\usage{
rec_song(
  genre,
  mode = NULL,
  energy = NULL,
  valence = NULL,
  danceability = NULL,
  instrumentalness = NULL,
  p = TRUE
)
}
\arguments{
\item{genre}{the genre}

\item{mode}{the mode, 0 for minor 1 for major}

\item{energy}{the energy of the song on a scale from 0.0-1.0}

\item{valence}{describes the musical positiveness of a song on a scale from
0.0-1.0. Higher valence songs are more happy/euphoric, lower valence songs
are more angry/depressed.}

\item{danceability}{describes how suitable the song is for dancing on a scale
from 0.0-1.0.}

\item{instrumentalness}{predicts if a song has no vocals on a scale from
0.0-1.0. The higher the instrumentalness the greater the chance the song
contains no vocal content.}

\item{p}{Whether the song will be printed (mostly internal use when building
playlists)}
}
\value{
an object of the song class with the following elements
\itemize{
\item song
\item author
\item duration
\item genre
\item danceability
\item energy
\item tempo
}
}
\description{
Given a genre of song and optionally other
}
\examples{
rec_song("pop", mode = 1, energy = 0.7, valence = 0.5, danceability = 0.8,
instrumentalness = 0.7)

}
