% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reccomendor.R
\name{rec_song}
\alias{rec_song}
\title{Recommend a song}
\usage{
rec_song(
  genre,
  mode = NULL,
  energy = NULL,
  loudness = NULL,
  valence = NULL,
  danceability = NULL,
  instrumentalness = NULL,
  min_duration = NULL,
  max_duration = NULL
)
}
\description{
Given a song title or artist (or other items tbd..)
}
