% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{summary.playlist}
\alias{summary.playlist}
\title{Summarizing Songs in the playlist}
\usage{
\method{summary}{playlist}(object, ...)
}
\arguments{
\item{object}{An object of the type playlist to be summarized}

\item{...}{to adhere to generic summary method}
}
\value{
A data frame including the Title, Genre, Author, Danceability,
Energy, Key and Tempo of each song in the playlist.
}
\description{
Summarizes each song & it's characteristics in the form of a
data frame to be easily analyzed or combined with other song data.
}
