% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Playlist_Class.R
\name{summary.playlist}
\alias{summary.playlist}
\title{Summarizing Songs in the playlist}
\usage{
\method{summary}{playlist}(x)
}
\arguments{
\item{obj}{An object of the type playlist}
}
\description{
Summarizes each song & it's characteristics
}
