% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{summary.song}
\alias{summary.song}
\title{Song Summary}
\usage{
\method{summary}{song}(song)
}
\arguments{
\item{song}{an object of type song to summarize in more detail}
}
\description{
Presents information about song in table forms
}
