% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Song_Class.R
\name{summary.song}
\alias{summary.song}
\title{Song Summary}
\usage{
\method{summary}{song}(object, ...)
}
\arguments{
\item{object}{an object of type song to summarize in more detail}

\item{...}{to adhere to generic summary method}
}
\description{
Presents information about song in table forms
}
